Finish autograder for p2, merged with p1_autograder
Now the autograder is renamed to autograder.sh and can be used to grader
both p1 and p2.
Usage:
  ./autograder.sh px [studentDir]
  where px=p1 | p2

prepare.sh is also changed to work for both p1 and p2.
Usage:
  ./prepare.sh px
It
- dtrx submissions_px.zip
- mv submissions_px px_gradingDir
- ./process_px_mkdir.sh
- ./process_px_copy.sh

p1 and p2 use the same test input. While new samples are generated for
p2 using lexer.

makefile is changed to include command to compile lexer
- If using clang, -fPIC needs to be added
- If using cc, no extra flags are needed